EST-1017: moves PostalAddressDtoFactory under org.estatio.dom
